Update default_session_token.js
Some applications may want to access the link hash, which can be returned by the creator.
Apps that do not manage the lifecycle of a session only need to map the session to the link hash. Two new methods for quick entry and removal in the table `_session_checking_tokens`.